Renamed primary router folder/prefix from `grid` to `grid3d`
