+++ frontend/src/api/ApiService.ts
@@ -8,7 +8,7 @@ import { AxiosHttpRequest } from './core/AxiosHttpRequest';
 import { DefaultService } from './services/DefaultService';
 import { ExploreService } from './services/ExploreService';
 import { GraphService } from './services/GraphService';
-import { GridService } from './services/GridService';
+import { Grid3DService } from './services/Grid3DService';
 import { InplaceVolumetricsService } from './services/InplaceVolumetricsService';
 import { ObservationsService } from './services/ObservationsService';
 import { ParametersService } from './services/ParametersService';
@@ -25,7 +25,7 @@ export class ApiService {
     public readonly default: DefaultService;
     public readonly explore: ExploreService;
     public readonly graph: GraphService;
-    public readonly grid: GridService;
+    public readonly grid3D: Grid3DService;
     public readonly inplaceVolumetrics: InplaceVolumetricsService;
     public readonly observations: ObservationsService;
     public readonly parameters: ParametersService;
@@ -53,7 +53,7 @@ export class ApiService {
         this.default = new DefaultService(this.request);
         this.explore = new ExploreService(this.request);
         this.graph = new GraphService(this.request);
-        this.grid = new GridService(this.request);
+        this.grid3D = new Grid3DService(this.request);
         this.inplaceVolumetrics = new InplaceVolumetricsService(this.request);
         this.observations = new ObservationsService(this.request);
         this.parameters = new ParametersService(this.request);
