+++ frontend/src/modules/Grid3DIntersection/queryHooks.ts
@@ -16,7 +16,7 @@ export function useGridIntersection(
     return useQuery({
         queryKey: ["gridParameterIntersection", caseUuid, ensembleName, gridName, parameterName, realization],
         queryFn: () =>
-            apiService.grid.gridParameterIntersection(
+            apiService.grid3D.gridParameterIntersection(
                 caseUuid ?? "",
                 ensembleName ?? "",
                 gridName ?? "",
@@ -46,7 +46,7 @@ export function useStatisticalGridIntersection(
             realizations,
         ],
         queryFn: () =>
-            apiService.grid.statisticalGridParameterIntersection(
+            apiService.grid3D.statisticalGridParameterIntersection(
                 caseUuid ?? "",
                 ensembleName ?? "",
                 gridName ?? "",
@@ -62,7 +62,7 @@ export function useStatisticalGridIntersection(
 export function useGridModelNames(caseUuid: string | null, ensembleName: string | null): UseQueryResult<string[]> {
     return useQuery({
         queryKey: ["getGridModelNames", caseUuid, ensembleName],
-        queryFn: () => apiService.grid.getGridModelNames(caseUuid ?? "", ensembleName ?? ""),
+        queryFn: () => apiService.grid3D.getGridModelNames(caseUuid ?? "", ensembleName ?? ""),
         staleTime: STALE_TIME,
         gcTime: CACHE_TIME,
         enabled: caseUuid && ensembleName ? true : false,
@@ -76,7 +76,7 @@ export function useGridParameterNames(
 ): UseQueryResult<string[]> {
     return useQuery({
         queryKey: ["getParameterNames", caseUuid, ensembleName, gridName],
-        queryFn: () => apiService.grid.getParameterNames(caseUuid ?? "", ensembleName ?? "", gridName ?? ""),
+        queryFn: () => apiService.grid3D.getParameterNames(caseUuid ?? "", ensembleName ?? "", gridName ?? ""),
         staleTime: STALE_TIME,
         gcTime: CACHE_TIME,
         enabled: caseUuid && ensembleName && gridName ? true : false,
